Patch forest to resolve the jumping trees problem.
The culprit was `\pgfsyspdfmark {pgfidNNN}{...}{...}` in the `.aux` file.  When
switching from the compilation of one chapter to another, sometimes the
tikzpicture ID (pgfidNNN) was just wrong enough to confuse Forest's anchor
computation.  Now Forest simply undefines the internal macro
\pgf@sys@pdf@mark@pos@pgfidNNN.  This should be ok, as there is simply no way
that the tikzpicture Forest uses internally to compute the anchor could have
gotten the mark legally.